+++ CHANGELOG.md
@@ -5,10 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Added lower bound validator to `freqs` in mode solver.
+
 ### Fixed
 - Bug in `Simulation.eps_bounds` that was always setting the lower bound to 1.
 - Extended lower limit of frequency range for `Graphene` to zero.
 - Improved warnings for `Medium2D`.
+- Improved mode solver handling of 1D problem to avoid singular matrix issue.
 
 ## [2.4.0rc2] - 2023-8-21